+++ setup_moves/ft_moves.h
@@ -6,7 +6,7 @@
 /*   By: tbouder <tbouder@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/05/10 22:10:56 by tbouder           #+#    #+#             */
-/*   Updated: 2016/05/18 18:41:13 by tbouder          ###   ########.fr       */
+/*   Updated: 2016/05/19 16:00:27 by tbouder          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 # include "../ft_lemin.h"
 
-int		*ft_init_tab(int ant);
-void	ft_print_moves(t_env *env, int *nbr, char **str);
+ULL		*ft_init_tab(ULL ant);
+void	ft_print_moves(t_env *env, ULL *nbr, char **str);
 
 int		*ft_init_mult_tab(int ant, int len);
 void	ft_print_mult_moves(t_env *env, int *nbr, char ***str);
